+++ src/mortimer/forms.py
@@ -124,8 +124,7 @@ class LocalExperimentForm(FlaskForm):
                 "You already have a local experiment with this title. Please choose a unique title.")
 
     def validate_exp_id(self, exp_id):
-        experiment = LocalExperiment.objects(exp_id__exact=exp_id.data).first()
-        print(experiment)
+        experiment = LocalExperiment.objects(exp_id__exact=exp_id.data, author__exact=current_user.username).first()
         if experiment is not None:
 
             raise ValidationError('There already exists an experiment with this ID in the database. Please choose a unique ID. If you already collected data using this ID, please contanct an administrator.')